pyproject.toml
[project]
name = "telegramaibot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "telegramify-markdown[mermaid]>=0.4.0",
    "watchdog>=2.1.6",  # Added watchdog dependency
]
